+++ app/jobs/record_submitted_claim_job.rb
@@ -1,0 +1,7 @@
+require "geckoboard"
+
+class RecordSubmittedClaimJob < ApplicationJob
+  def perform(claim)
+    Claim::GeckoboardEvent.new(claim, :submitted, claim.submitted_at).record
+  end
+end